feat(byline): add author formatting
