+++ site/clientFormatting.tsx
@@ -1,0 +1,20 @@
+import { last } from "@ourworldindata/utils"
+
+export const formatAuthors = ({
+    authors,
+    requireMax,
+    forBibtex,
+}: {
+    authors: string[]
+    requireMax?: boolean
+    forBibtex?: boolean
+}) => {
+    if (requireMax && !authors.includes("Max Roser"))
+        authors = [...authors, "Max Roser"]
+
+    let authorsText = authors.slice(0, -1).join(forBibtex ? " and " : ", ")
+    if (authorsText.length === 0) authorsText = authors[0]
+    else authorsText += ` and ${last(authors)}`
+
+    return authorsText
+}
