+++ site/formatting.tsx
@@ -8,7 +8,6 @@ import {
     WP_BlockClass,
     WP_ColumnStyle,
     WP_PostType,
-    last,
     formatDate,
 } from "@ourworldindata/utils"
 import { BAKED_BASE_URL, WORDPRESS_URL } from "../settings/serverSettings.js"
@@ -35,25 +34,6 @@ export const formatUrls = (html: string) =>
         .replace(new RegExp("https?://owid.cloud", "g"), BAKED_BASE_URL)
         .replace(new RegExp("https?://ourworldindata.org", "g"), BAKED_BASE_URL)
         .replace(new RegExp("/app/uploads", "g"), "/uploads")
-
-export const formatAuthors = ({
-    authors,
-    requireMax,
-    forBibtex,
-}: {
-    authors: string[]
-    requireMax?: boolean
-    forBibtex?: boolean
-}) => {
-    if (requireMax && !authors.includes("Max Roser"))
-        authors = [...authors, "Max Roser"]
-
-    let authorsText = authors.slice(0, -1).join(forBibtex ? " and " : ", ")
-    if (authorsText.length === 0) authorsText = authors[0]
-    else authorsText += ` and ${last(authors)}`
-
-    return authorsText
-}
 
 export const splitContentIntoSectionsAndColumns = (
     cheerioEl: CheerioStatic
