+++ site/gdocs/centered-article.scss
@@ -189,6 +189,7 @@
 
     > div {
         width: 50%;
+        padding: 0 16px;
 
         &:first-child {
             border-right: 1px solid $blue-10;
